feat/made instructions work for all games
